+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - 新增：schema BaseSchema 参数 `deserialize_postprocess`
 - 新增：schema Datetime 参数 `with_timezone`
+- 新增：schema default 参数可以使用函数
 - 修复一些BUG
 
 ## 0.1a3